modified:   Rakefile modified:   lib/match.rb new file:   lib/matches_history.rb modified:   lib/player.rb modified:   lib/user.rb modified:   lib/valorant_api.rb new file:   spec/helpers/deathmatch.json modified:   spec/helpers/match.json new file:   spec/helpers/matches_history.json modified:   spec/match_spec.rb new file:   spec/matches_history_spec.rb modified:   spec/player_spec.rb
Implemented a new system for the rake. Now it will be possible to run the rake with the following commands:
rake history - will run the rake for the matches history
rake match - will run the rake for the match
rake player - will run the rake for the player
rake user - will run the rake for the user

Implemented the matches history class. It will be possible to get the matches history of a player.
Changed behavior in match class for non-team-based game modes.